+++ DiamondDataset2.py
@@ -51,7 +51,7 @@ class DiamondDataset2(_DiamondDataset, ABC):
         out = super()._preprocess_labels(labels)
         assert out[1] is None
 
-        out[1] = torch.zeros(len(self.shape_labels))
+        out[1] = torch.zeros(len(self.shape_labels), dtype=torch.double)
         out[1][self.shape_labels.index(labels[1])] = 1
 
         return out
